+++ test/nextjournal/clerk/builder_test.clj
@@ -58,6 +58,10 @@
            (builder/expand-paths {:paths ["notebooks/rule_**.clj"]
                                   :index (str (fs/file "notebooks" "rule_30.clj"))}))))
 
+  (testing "supports absolute paths (#504)"
+    (is (= {:expanded-paths [(str (fs/file (fs/cwd) "book.clj"))]}
+           (builder/expand-paths {:paths [(str (fs/file (fs/cwd) "book.clj"))]}))))
+
   (testing "invalid args"
     (is (match? {:error #"must set either"}
                 (builder/expand-paths {})))
@@ -99,13 +103,26 @@
 
 (deftest doc-url
   (testing "link to same dir unbundled"
-    (is (= "./../notebooks/rule_30.html" ;; NOTE: could also be just "rule_30.html"
-           (builder/doc-url {:bundle? false} [{:file "notebooks/viewer_api.clj"} {:file "notebooks/rule_30.clj"}] "notebooks/viewer_api.clj" "notebooks/rule_30.clj"))))
+    ;; in the unbundled case the current URL on a given notebook is given by
+    ;;
+    ;; fs-path              |  URL
+    ;; ----------------------------------------------------
+    ;; path/to/notebook.clj |  path/to/notebok/[index.html]
+    (is (= "./../../notebooks/rule_30" ;; NOTE: could also be just "rule_30.html"
+           (builder/doc-url {:bundle? false} "notebooks/viewer_api.clj" "notebooks/rule_30"))))
 
   (testing "respects the mapped index"
-    (is (= "./notebooks/rule_30.html"
-           (builder/doc-url {:bundle? false} [{:file "index.clj"} {:file "notebooks/rule_30.clj"}] "index.clj" "notebooks/rule_30.clj"))))
+    (is (= "./notebooks/rule_30"
+           (builder/doc-url {:bundle? false} "index.clj" "notebooks/rule_30")))
+
+    (is (= "./notebooks/rule_30"
+           (builder/doc-url {:bundle? false :index "notebooks/path/to/notebook.clj"}
+                            "notebooks/path/to/notebook.clj" "notebooks/rule_30")))
+
+    (is (= "./../../../../notebooks/rule_30"
+           (builder/doc-url {:bundle? false}
+                            "notebooks/path/to/notebook.clj" "notebooks/rule_30"))))
 
   (testing "bundle case"
-    (is (= "#/notebooks/rule_30.clj"
-           (builder/doc-url {:bundle? true} [{:file "notebooks/index.clj"} {:file "notebooks/rule_30.clj"}] "index.clj" "notebooks/rule_30.clj")))))
+    (is (= "#/notebooks/rule_30"
+           (builder/doc-url {:bundle? true} "index.clj" "notebooks/rule_30")))))